Refactor RuleViolation into an interface
